Fixed compiler warning.

+++ src/rf_phreaker/fir_filter/fir_filter_impl.cpp
@@ -57,10 +57,10 @@ fir_filter_impl::~fir_filter_impl()
 void fir_filter_impl::set_up_down_factor_based_on_sampling_rates(frequency_type original_sampling_rate, frequency_type new_sampling_rate)
 {
 	if(original_sampling_rate % new_sampling_rate == 0) {
-		set_up_down_factor(1, original_sampling_rate / new_sampling_rate);
+		set_up_down_factor(1, (int)(original_sampling_rate / new_sampling_rate));
 	}
 	else if(new_sampling_rate % original_sampling_rate == 0) {
-		set_up_down_factor(new_sampling_rate / original_sampling_rate, 1);
+		set_up_down_factor((int)(new_sampling_rate / original_sampling_rate), 1);
 	}
 	else {
 		auto lcm = boost::math::lcm(original_sampling_rate, new_sampling_rate);
